refactor: remove PokemonAPI class
- replace class with standalone functions
- add tests for new functions
